+++ lib/onlyoffice_webdriver_wrapper/webdriver/webdriver_js_methods.rb
@@ -60,5 +60,17 @@ module OnlyofficeWebdriverWrapper
     def document_ready?
       execute_javascript('return document.readyState;') == 'complete'
     end
+
+    # Get ComputedStyle property value
+    # @param xpath [String] xpath of element
+    # @param pseudo_element [String] pseudo element to compute
+    # @param property [String] property to get
+    # @return [String] value of property
+    def computed_style(xpath, pseudo_element = 'null', property = nil)
+      element_by_xpath = "document.evaluate(\"#{xpath.tr('"', "'")}\",document, null, XPathResult.ANY_TYPE, null ).iterateNext()"
+      style = "window.getComputedStyle(#{element_by_xpath}, '#{pseudo_element}')"
+      result = "#{style}.getPropertyValue('#{property}')"
+      execute_javascript("return #{result}")
+    end
   end
 end